fix(monorepo): resolve JPMS split-package via Option B1 (relocate 12 survivors to forge21.*, lift SkinUtils to :common)
Relocate MC-bound survivors out of the 5 packages shared with :common
(util, permission, skinchanger, metrics, integration.discordsrv) into
version-scoped levosilimo.everlastingskins.forge21.* so the dev-run
module path no longer sees split packages with the common automatic
module. Lift SkinUtils into :common (adapted to the Java 8 --release
gate: Stream.toList -> Collectors.toList, Optional.isEmpty ->
!isPresent). JavaHttpClient, VanillaPermissionService and
LuckPermsPermissionService stay forge-side: they are transitively
MC-bound (Config/ForgeConfigSpec) or JDK-11-only (java.net.http).

Also drop the config read fix-1 added to the EverlastingSkins
constructor (ForgeConfigSpec.get() before registerConfig throws);
DiscordSrvConfig is configured from onServerAboutToStart, where the
config is loaded.

Gate: :common:build, :forge-1.21:test (215/215), runGameTestServer
(34/34, no ResolutionException).
