+++ common/src/main/java/levosilimo/everlastingskins/util/SkinUtils.java
@@ -18,6 +18,7 @@ import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.*;
 import java.util.concurrent.ThreadLocalRandom;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class SkinUtils {
@@ -62,7 +63,7 @@ public class SkinUtils {
     public static void renameFile(Path parent, String oldName, String newName) throws IOException {
         try (Stream<Path> stream = Files.list(parent)) {
             // Folders are case-insensitive on Windows, so we need to check it using this method
-            List<String> files = stream.map(Path::getFileName).map(Path::toString).toList();
+            List<String> files = stream.map(Path::getFileName).map(Path::toString).collect(Collectors.toList());
 
             String tempName = newName + "_temp";
             if (files.contains(oldName) && !files.contains(tempName) && !files.contains(newName)) {
@@ -141,7 +142,7 @@ public class SkinUtils {
 
     public static String sanitizeImageURL(String imageUrl) {
         Optional<URL> uriOptional = parseURL(imageUrl);
-        if (uriOptional.isEmpty()) {
+        if (!uriOptional.isPresent()) {
             return imageUrl;
         }
 
@@ -172,7 +173,7 @@ public class SkinUtils {
 
     public static String sanitizeSkinInput(String skinInput) {
         Optional<URL> uriOptional = parseURL(skinInput);
-        if (uriOptional.isEmpty()) {
+        if (!uriOptional.isPresent()) {
             return skinInput;
         }
