+++ forge-1.21/src/main/java/levosilimo/everlastingskins/broadcast/VanillaSkinBroadcaster.java
@@ -10,9 +10,9 @@ import com.mojang.authlib.GameProfile;
 import io.netty.buffer.Unpooled;
 import levosilimo.everlastingskins.Config;
 import levosilimo.everlastingskins.EverlastingSkins;
-import levosilimo.everlastingskins.metrics.NetworkMetricsHandler;
+import levosilimo.everlastingskins.forge21.metrics.NetworkMetricsHandler;
 import levosilimo.everlastingskins.metrics.SkinMetrics;
-import levosilimo.everlastingskins.skinchanger.SkinRestorer;
+import levosilimo.everlastingskins.forge21.skinchanger.SkinRestorer;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.protocol.Packet;
 import net.minecraft.network.protocol.game.ClientGamePacketListener;
