+++ forge-1.21/src/main/java/levosilimo/everlastingskins/forge21/permission/PermissionContext.java
@@ -5,7 +5,7 @@
  */
 
 
-package levosilimo.everlastingskins.permission;
+package levosilimo.everlastingskins.forge21.permission;
 
 import net.minecraft.server.level.ServerPlayer;
