+++ forge-1.21/src/main/java/levosilimo/everlastingskins/forge21/util/CompletionSources.java
@@ -4,10 +4,10 @@
  * https://github.com/Levosilimo/Everlasting-Skins
  */
 
-package levosilimo.everlastingskins.util;
+package levosilimo.everlastingskins.forge21.util;
 
 import levosilimo.everlastingskins.Config;
-import levosilimo.everlastingskins.permission.PermissionContext;
+import levosilimo.everlastingskins.forge21.permission.PermissionContext;
 import levosilimo.everlastingskins.permission.PermissionServiceManager;
 import levosilimo.everlastingskins.skinchanger.DefaultSkinResolver;
 import levosilimo.everlastingskins.skinchanger.MojangProfileCache;
