+++ forge-1.21/src/test/java/levosilimo/everlastingskins/integration/placeholderapi/EverlastingSkinsExpansionTest.java
@@ -7,7 +7,7 @@
 
 package levosilimo.everlastingskins.integration.placeholderapi;
 
-import levosilimo.everlastingskins.skinchanger.SkinRestorer;
+import levosilimo.everlastingskins.forge21.skinchanger.SkinRestorer;
 import levosilimo.everlastingskins.skinchanger.SkinStorage;
 import org.bukkit.OfflinePlayer;
 import org.junit.jupiter.api.AfterEach;
