+++ forge-1.21/src/test/java/levosilimo/everlastingskins/permission/PermissionServiceManagerTest.java
@@ -7,6 +7,8 @@
 
 package levosilimo.everlastingskins.permission;
 
+import levosilimo.everlastingskins.forge21.permission.PermissionContext;
+import levosilimo.everlastingskins.forge21.permission.VanillaPermissionService;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
