+++ forge-1.21/src/test/java/levosilimo/everlastingskins/permission/VanillaPermissionServiceTest.java
@@ -7,6 +7,8 @@
 
 package levosilimo.everlastingskins.permission;
 
+import levosilimo.everlastingskins.forge21.permission.PermissionContext;
+import levosilimo.everlastingskins.forge21.permission.VanillaPermissionService;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
