+++ forge-1.21/src/test/java/levosilimo/everlastingskins/skinchanger/SkinActionStoredSourceSkipTest.java
@@ -3,6 +3,7 @@
  */
 package levosilimo.everlastingskins.skinchanger;
 
+import levosilimo.everlastingskins.forge21.skinchanger.SkinRestorer;
 import levosilimo.everlastingskins.enums.SkinVariant;
 import levosilimo.everlastingskins.permission.TestConfigSupport;
 import levosilimo.everlastingskins.skinchanger.command.SkinActionCommand;
