+++ forge-1.21/src/test/java/levosilimo/everlastingskins/skinchanger/SkinCommandTabCompleteTest.java
@@ -11,11 +11,12 @@ import com.mojang.brigadier.ParseResults;
 import com.mojang.brigadier.suggestion.Suggestion;
 import com.mojang.brigadier.suggestion.Suggestions;
 import com.mojang.brigadier.tree.CommandNode;
+import levosilimo.everlastingskins.forge21.skinchanger.SkinCommand;
 import levosilimo.everlastingskins.Config;
 import levosilimo.everlastingskins.permission.PermissionServiceManager;
 import levosilimo.everlastingskins.permission.TestConfigSupport;
-import levosilimo.everlastingskins.permission.VanillaPermissionService;
-import levosilimo.everlastingskins.util.CompletionSources;
+import levosilimo.everlastingskins.forge21.permission.VanillaPermissionService;
+import levosilimo.everlastingskins.forge21.util.CompletionSources;
 import levosilimo.everlastingskins.util.CustomSkinProperty;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.server.Bootstrap;
